Move BarConfig to its own class
